Add Ubuntu to prerequisites section
I added Ubuntu commands for installing prerequisites.

The difference between h3 and h4 is hard to spot visually. One effect of this is that it looked a little bit like Mac was a prerequisite, when Mac was intended to be an informative section inside Prerequisites. I tried to make this more clear. I also tried to make the dependency list more clear. Before it looked like the dependencies were Mac, Erlang, Mac, Elixir, etc. Now the dependency list should be clear with platform-specific instructions underneath.